refactor: split install pypi into module and files
